refactor: move types dir into react
