+++ src/react/types/error/BaseError.ts
